+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2020-07-16
+* Implement Effective Dart changes
+* Add a safety net related to using the fader controller after it's been disposed
+
 ## [1.0.2+1] - 2020-07-16
 * Add automated tests w/ code coverage